+++ src/EllipticalCopulas/GaussianCopula.jl
@@ -78,7 +78,7 @@ end
 ρ(C::GaussianCopula{2,MT}) where MT = 6*asin(C.Σ[1,2]/2)/π
 
 # Conditioning and subsetting fast paths colocated with the type
-@inline function DistortionFromCop(C::GaussianCopula{D,MT}, js::NTuple{p,Int}, uⱼₛ::NTuple{p,Float64}, i::Int64) where {D,MT,p}
+@inline function DistortionFromCop(C::GaussianCopula{D,MT}, js::NTuple{p,Int64}, uⱼₛ::NTuple{p,Float64}, i::Int64) where {D,MT,p}
     ist = Tuple(setdiff(1:D, js))
     @assert i in ist
     J = collect(js)
@@ -95,7 +95,7 @@ end
 end
 @inline function ConditionalCopula(C::GaussianCopula{D,MT}, js::NTuple{p,Int64}, uⱼₛ::NTuple{p,Float64}) where {D,MT,p}
     @assert 0 < p < D-1
-    J = collect(Int, js)
+    J = collect(Int64, js)
     I = collect(setdiff(1:D, J))
     Σcond = C.Σ[I, I] - C.Σ[I, J] * inv(C.Σ[J, J]) * C.Σ[J, I]
     return GaussianCopula(Σcond)
