+++ src/UnivariateDistribution/Distortions/BivFGMDistortion.jl
@@ -15,4 +15,4 @@ end
     # Solve a u^2 - (1+a) u + α = 0 and pick the root in [0,1]
     return ((1 + a) - sqrt((1 + a)^2 - 4*a*α)) / 2a
 end
-@inline DistortionFromCop(C::FGMCopula{2}, js::NTuple{1,Int}, uⱼₛ::NTuple{1,Float64}, ::Int64) = BivFGMDistortion(float(C.θ[1]), Int8(js[1]), float(uⱼₛ[1]))
+@inline DistortionFromCop(C::FGMCopula{2}, js::NTuple{1,Int64}, uⱼₛ::NTuple{1,Float64}, ::Int64) = BivFGMDistortion(float(C.θ[1]), Int8(js[1]), float(uⱼₛ[1]))
